капитан ООП
