+++ src/data.js
@@ -1,3 +1,5 @@
+import {rand} from './random.js';
+
 const data = () => ({
   routeType: new Map([
     [`Taxi`, `🚕`],
@@ -20,7 +22,7 @@ const data = () => ({
     `Dacca`,
     `Minsk`,
   ],
-  photo: `http://picsum.photos/300/150?r=${Math.random()}`,
+  photo: `http://picsum.photos/300/150?r=`,
   offers: new Set([
     `Add luggage`,
     `Switch to comfort class`,
@@ -37,7 +39,7 @@ const data = () => ({
   Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`,
   dateTrip: Date.now() + 1 + Math.floor(Math.random() * 7 * 24 * 60) * 60 * 1000,
   price: Math.floor(Math.random() * 15 * 10),
-  durationPerMinutes: Math.floor(Math.random() * (10000000 - 10000) + 10000),
+  durationPerMinutes: rand(10000000, 10000),
 });
 
 export default data;
