+++ src/set-current-data.js
@@ -4,7 +4,7 @@ import {rand} from './random.js';
 const currentData = (data) => ({
   routeType: generType(data.routeType),
   cities: data.cities[rand(7, 0)],
-  photo: data.photo,
+  photo: data.photo + `${Math.random()}`,
   offers: setPriceToOffers(getRandomOffers(data.offers)),
   text: getText(data.text),
   dateTrip: data.dateTrip,
